create ProgrammableDIOClass as separate module
